+++ examples/create_customer.rb
@@ -1,19 +1,19 @@
 # frozen_string_literal: true
 
-require 'tuneuptechnology'
+require_relative 'tuneuptechnology'
 
-data = {
-  'auth' => ENV['API_EMAIL'],
-  'api_key' => ENV['API_KEY'],
-  'firstname' => 'Jake',
-  'lastname' => 'Peralta',
-  'email' => 'jake@example.com',
-  'phone' => '8015551234',
-  'user_id' => 1,
-  'notes' => 'Believes he is a good detective.',
-  'location_id' => 1
-}
+client = TuneupTechnology::Client.new ENV['API_EMAIL'], ENV['API_KEY']
 
-customer = TuneupTechnology::Customer.create(data)
+customer = client.customers.create(
+  {
+    'firstname' => 'Jake',
+    'lastname' => 'Peralta',
+    'email' => 'jake@example.com',
+    'phone' => '8015551234',
+    'user_id' => 1,
+    'notes' => 'Believes he is a good detective.',
+    'location_id' => 2
+  }
+)
 
 puts customer