+++ examples/update_customer.rb
@@ -2,19 +2,18 @@
 
 require 'tuneuptechnology'
 
-data = {
-  'auth' => ENV['API_EMAIL'],
-  'api_key' => ENV['API_KEY'],
-  'id' => 1, # the ID of the customer you are updating
-  'firstname' => 'Jake',
-  'lastname' => 'Peralta',
-  'email' => 'jake@example.com',
-  'phone' => '8015551234',
-  'user_id' => 1,
-  'notes' => 'Believes he is a good detective.',
-  'location_id' => 1
-}
+client = TuneupTechnology::Client.new ENV['API_EMAIL'], ENV['API_KEY']
 
-customer = TuneupTechnology::Customer.create(data)
+customer = client.customers.update(
+  {
+    'firstname' => 'Jake',
+    'lastname' => 'Peralta',
+    'email' => 'jake@example.com',
+    'phone' => '8015551234',
+    'user_id' => 1,
+    'notes' => 'Believes he is a good detective.',
+    'location_id' => 2
+  }
+)
 
 puts customer