+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## v2.0.0 (TODO)
+
+* Updates entire library to be compliant with the new `v2` API (endpoints, HTTP methods, etc)
+* Proper error handling by wrapping the RestClient request in begin/rescue
+* The return value of any method is now the `Response` object instead of `json`, now you can retrieve the `status_code` or other details of the response
+* Build requests via a `Client` now providing your email and api_key
+* Added optional `base_url` and `timeout` options to client
+* Module names are now plural
+* The Client now checks if an email and api_key is provided and raises an error if not
+* Added unit tests (closes #1)
+
 ## v1.2.0 (2021-02-20)
 
 * Adds a timeout for HTTP requests